[CSPM] refactor compliance load command into pkg/compliance/cli (#43223)
### What does this PR do?

Similar to https://github.com/DataDog/datadog-agent/pull/43141 but this time for the `compliance load` command. This will help sharing this command code when the compliance CLI is available in the system-probe.

### Motivation

### Describe how you validated your changes

### Additional Notes
